Time: 1 ms (97.35%), Space: 42.4 MB (89.41%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -8,7 +8,7 @@ class Solution {
         while (low <= high) {
             int mid = low + (high - low) / 2;
 
-            if ( mid * mid <= x) {
+            if ((long) mid * mid <= x) {
                 ans = mid;          
                 low = mid + 1;     
             } else {
